+++ runners/google-cloud-dataflow-java/worker/src/test/java/org/apache/beam/runners/dataflow/worker/StreamingGroupAlsoByWindowFnsTest.java
@@ -149,6 +149,7 @@ public class StreamingGroupAlsoByWindowFnsTest {
                 TimerData.of(
                     namespace,
                     timestamp,
+                    timestamp,
                     type == Windmill.Timer.Type.WATERMARK
                         ? TimeDomain.EVENT_TIME
                         : TimeDomain.PROCESSING_TIME)))