+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/control/DefaultJobBundleFactory.java
@@ -20,6 +20,7 @@ package org.apache.beam.runners.fnexecution.control;
 import com.google.auto.value.AutoValue;
 import java.io.IOException;
 import java.util.IdentityHashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
@@ -42,6 +43,7 @@ import org.apache.beam.runners.core.construction.graph.ExecutableStage;
 import org.apache.beam.runners.fnexecution.GrpcContextHeaderAccessorProvider;
 import org.apache.beam.runners.fnexecution.GrpcFnServer;
 import org.apache.beam.runners.fnexecution.ServerFactory;
+import org.apache.beam.runners.fnexecution.artifact.ArtifactRetrievalService;
 import org.apache.beam.runners.fnexecution.artifact.BeamFileSystemLegacyArtifactRetrievalService;
 import org.apache.beam.runners.fnexecution.artifact.ClassLoaderLegacyArtifactRetrievalService;
 import org.apache.beam.runners.fnexecution.artifact.LegacyArtifactRetrievalService;
@@ -234,6 +236,11 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
                       environmentFactoryProviderMap.get(environment.getUrn());
                   ServerFactory serverFactory = environmentFactoryProvider.getServerFactory();
                   ServerInfo serverInfo = serverInfoCreator.apply(serverFactory);
+                  String workerId = stageIdGenerator.getId();
+                  serverInfo
+                      .getProvisioningServer()
+                      .getService()
+                      .registerEnvironment(workerId, environment);
                   EnvironmentFactory environmentFactory =
                       environmentFactoryProvider.createEnvironmentFactory(
                           serverInfo.getControlServer(),
@@ -243,7 +250,7 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
                           clientPool,
                           stageIdGenerator);
                   return WrappedSdkHarnessClient.wrapping(
-                      environmentFactory.createEnvironment(environment), serverInfo);
+                      environmentFactory.createEnvironment(environment, workerId), serverInfo);
                 }
               });
 
@@ -613,6 +620,7 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
       // These will be closed in the reverse creation order:
       try (AutoCloseable envCloser = environment;
           AutoCloseable provisioningServer = serverInfo.getProvisioningServer();
+          AutoCloseable legacyRetrievalServer = serverInfo.getLegacyRetrievalServer();
           AutoCloseable retrievalServer = serverInfo.getRetrievalServer();
           AutoCloseable stateServer = serverInfo.getStateServer();
           AutoCloseable dataServer = serverInfo.getDataServer();
@@ -653,12 +661,12 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
     PortablePipelineOptions portableOptions =
         PipelineOptionsTranslation.fromProto(jobInfo.pipelineOptions())
             .as(PortablePipelineOptions.class);
-    LegacyArtifactRetrievalService artifactRetrievalService;
+    LegacyArtifactRetrievalService legacyArtifactRetrievalService;
 
     if (portableOptions.getRetrievalServiceType() == RetrievalServiceType.CLASSLOADER) {
-      artifactRetrievalService = new ClassLoaderLegacyArtifactRetrievalService();
+      legacyArtifactRetrievalService = new ClassLoaderLegacyArtifactRetrievalService();
     } else {
-      artifactRetrievalService = BeamFileSystemLegacyArtifactRetrievalService.create();
+      legacyArtifactRetrievalService = BeamFileSystemLegacyArtifactRetrievalService.create();
     }
 
     GrpcFnServer<FnApiControlClientPoolService> controlServer =
@@ -669,15 +677,23 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
     GrpcFnServer<GrpcLoggingService> loggingServer =
         GrpcFnServer.allocatePortAndCreateFor(
             GrpcLoggingService.forWriter(Slf4jLogWriter.getDefault()), serverFactory);
-    GrpcFnServer<LegacyArtifactRetrievalService> retrievalServer =
-        GrpcFnServer.allocatePortAndCreateFor(artifactRetrievalService, serverFactory);
+    List<GrpcFnServer<?>> artifactServers =
+        GrpcFnServer.allocatePortAndCreateFor(
+            ImmutableList.of(new ArtifactRetrievalService(), legacyArtifactRetrievalService),
+            serverFactory);
+    GrpcFnServer<ArtifactRetrievalService> retrievalServer =
+        (GrpcFnServer<ArtifactRetrievalService>) artifactServers.get(0);
+    GrpcFnServer<LegacyArtifactRetrievalService> legacyRetrievalServer =
+        (GrpcFnServer<LegacyArtifactRetrievalService>) artifactServers.get(1);
     ProvisionApi.ProvisionInfo.Builder provisionInfo = jobInfo.toProvisionInfo().toBuilder();
     provisionInfo.setLoggingEndpoint(loggingServer.getApiServiceDescriptor());
-    provisionInfo.setArtifactEndpoint(retrievalServer.getApiServiceDescriptor());
+    provisionInfo.setArtifactEndpoint(legacyRetrievalServer.getApiServiceDescriptor());
     provisionInfo.setControlEndpoint(controlServer.getApiServiceDescriptor());
     GrpcFnServer<StaticGrpcProvisionService> provisioningServer =
         GrpcFnServer.allocatePortAndCreateFor(
-            StaticGrpcProvisionService.create(provisionInfo.build()), serverFactory);
+            StaticGrpcProvisionService.create(
+                provisionInfo.build(), GrpcContextHeaderAccessorProvider.getHeaderAccessor()),
+            serverFactory);
     GrpcFnServer<GrpcDataService> dataServer =
         GrpcFnServer.allocatePortAndCreateFor(
             GrpcDataService.create(
@@ -691,6 +707,7 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
             .setControlServer(controlServer)
             .setLoggingServer(loggingServer)
             .setRetrievalServer(retrievalServer)
+            .setLegacyRetrievalServer(legacyRetrievalServer)
             .setProvisioningServer(provisioningServer)
             .setDataServer(dataServer)
             .setStateServer(stateServer)
@@ -705,7 +722,9 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
 
     abstract GrpcFnServer<GrpcLoggingService> getLoggingServer();
 
-    abstract GrpcFnServer<LegacyArtifactRetrievalService> getRetrievalServer();
+    abstract GrpcFnServer<ArtifactRetrievalService> getRetrievalServer();
+
+    abstract GrpcFnServer<LegacyArtifactRetrievalService> getLegacyRetrievalServer();
 
     abstract GrpcFnServer<StaticGrpcProvisionService> getProvisioningServer();
 
@@ -721,7 +740,10 @@ public class DefaultJobBundleFactory implements JobBundleFactory {
 
       abstract Builder setLoggingServer(GrpcFnServer<GrpcLoggingService> server);
 
-      abstract Builder setRetrievalServer(GrpcFnServer<LegacyArtifactRetrievalService> server);
+      abstract Builder setRetrievalServer(GrpcFnServer<ArtifactRetrievalService> server);
+
+      abstract Builder setLegacyRetrievalServer(
+          GrpcFnServer<LegacyArtifactRetrievalService> server);
 
       abstract Builder setProvisioningServer(GrpcFnServer<StaticGrpcProvisionService> server);