+++ sdks/go/pkg/beam/runners/direct/direct.go
@@ -239,17 +239,20 @@ func (b *builder) makeLink(id linkID) (exec.Node, error) {
 			Out:     out,
 			PID:     path.Base(edge.DoFn.Name()),
 		}
+		u = pardo
+		if edge.DoFn.IsSplittable() {
+			u = &exec.SdfFallback{PDo: pardo}
+		}
 		if len(edge.Input) == 1 {
-			u = pardo
 			break
 		}
 
 		// ParDo w/ side input. We need to insert buffering and wait. We also need to
 		// ensure that we return the correct link node.
 
-		b.units = append(b.units, pardo)
+		b.units = append(b.units, u)
 
-		w := &wait{UID: b.idgen.New(), need: len(edge.Input) - 1, next: pardo}
+		w := &wait{UID: b.idgen.New(), need: len(edge.Input) - 1, next: u}
 		b.units = append(b.units, w)
 		b.links[linkID{edge.ID(), 0}] = w