+++ sdks/python/apache_beam/testing/load_tests/pardo_test.py
@@ -93,11 +93,10 @@ class ParDoTest(LoadTest):
   def __init__(self):
     super(ParDoTest, self).__init__()
     self.iterations = self.get_option_or_default('iterations')
-    self.number_of_counters = self.get_option_or_default('number_of_counters')
-    self.number_of_operations = self.get_option_or_default(
-        'number_of_counter_operations', 1)
     self.number_of_counters = self.get_option_or_default(
         'number_of_counters', 1)
+    self.number_of_operations = self.get_option_or_default(
+        'number_of_counter_operations', 1)
     self.stateful = self.get_option_or_default('stateful', False)
     if self.get_option_or_default('state_cache', False):
       self.pipeline.options.view_as(DebugOptions).add_experiment(
@@ -175,33 +174,32 @@ class StatefulLoadGenerator(beam.PTransform):
 
   class GenerateLoad(beam.DoFn):
     state_spec = userstate.CombiningValueStateSpec(
-        'bundles_remaining', beam.coders.VarIntCoder(), sum)
+        'bundles_remaining', combine_fn=sum)
     timer_spec = userstate.TimerSpec('timer', userstate.TimeDomain.WATERMARK)
 
     def __init__(self, num_records_per_key, value_size, bundle_size=1000):
       self.num_records_per_key = num_records_per_key
       self.payload = os.urandom(value_size)
       self.bundle_size = bundle_size
-      self.key = None
 
     def process(
         self,
-        element,
+        _element,
         records_remaining=beam.DoFn.StateParam(state_spec),
         timer=beam.DoFn.TimerParam(timer_spec)):
-      self.key, _ = element
       records_remaining.add(self.num_records_per_key)
       timer.set(0)
 
     @userstate.on_timer(timer_spec)
     def process_timer(
         self,
+        key=beam.DoFn.KeyParam,
         records_remaining=beam.DoFn.StateParam(state_spec),
         timer=beam.DoFn.TimerParam(timer_spec)):
       cur_bundle_size = min(self.bundle_size, records_remaining.read())
       for _ in range(cur_bundle_size):
         records_remaining.add(-1)
-        yield self.key, self.payload
+        yield key, self.payload
       if records_remaining.read() > 0:
         timer.set(0)
 
@@ -213,11 +211,9 @@ class StatefulLoadGenerator(beam.PTransform):
         | 'Impulse' >> beam.Impulse()
         | 'GenerateKeys' >> beam.ParDo(
             StatefulLoadGenerator.GenerateKeys(self.num_keys, self.key_size))
-        | 'Reshuffle' >> beam.Reshuffle()
         | 'GenerateLoad' >> beam.ParDo(
             StatefulLoadGenerator.GenerateLoad(
-                self.num_records // self.num_keys, self.value_size))
-        | 'Reshuffle2' >> beam.Reshuffle())
+                self.num_records // self.num_keys, self.value_size)))
 
 
 if __name__ == '__main__':